requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Firebase Admin SDK
firebase-admin==6.5.0

# HTTP Requests (for 1.py script)
requests==2.32.3

# Environment Variables
python-dotenv==1.0.1

# Data Validation
pydantic==2.9.2
pydantic[email]==2.9.2

# CORS Support (included with FastAPI but explicit for clarity)
python-multipart==0.0.12

# Optional but recommended for production
gunicorn==23.0.0  # For production deployment

# Development/Testing (optional)
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2  # For testing FastAPI endpoints